requirements.txt
more-itertools==8.12.0
opencv-python==4.5.5.62
opensimplex==0.4
pycairo==1.20.1
